requirements.txt
defusedxml==0.4.1
docopt==0.6.2
dominate==2.1.12
dpath==1.4.0
Flask==0.10.1
Flask-Bootstrap==3.3.5.6
Flask-Login==0.2.11
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
oauthlib==1.0.3
passlib==1.6.5
git+https://github.com/wurstmineberg/people.git#egg=people
psycopg2==2.6.1
PyJWT==1.4.0
python-social-auth==0.2.12
python3-openid==3.0.6
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
SQLAlchemy==1.0.8
Werkzeug==0.10.4
